Droplet User-Data Fix

+++ src/vpndeployer/droplets.py
@@ -11,7 +11,13 @@ def create_droplet(api_token, ip, name, region, image, email):
         image=f'{image}',
         size_slug='512mb',
         user_data=f"""#!/bin/bash
-    yum -y install ansible
+    if [[ -e /etc/debian_version ]]; then
+        apt-get -y update
+        apt-get -y install ansible
+    else
+        yum -y update
+        yum -y install ansible
+    fi
     curl -o /root/openvpn-install.yml https://raw.githubusercontent.com/LopezNathan/vpn-deployer/development/openvpn-install.yml
     ansible-playbook /root/openvpn-install.yml --extra-vars "IP={ip} EMAIL={email}" > /var/log/ansible.log 2>&1
     """,
